move extractGIF1stFrame function to thumbnail/gif.go
